+++ source/utils/CarlaOscUtils.hpp
@@ -128,7 +128,7 @@ void osc_send_control(const CarlaOscData& oscData, const int32_t index, const fl
     char targetPath[std::strlen(oscData.path)+9];
     std::strcpy(targetPath, oscData.path);
     std::strcat(targetPath, "/control");
-    try_lo_send(oscData.target, targetPath, "if", index, value);
+    try_lo_send(oscData.target, targetPath, "if", index, static_cast<double>(value));
 }
 
 static inline
@@ -209,7 +209,7 @@ void osc_send_sample_rate(const CarlaOscData& oscData, const float sampleRate) n
     char targetPath[std::strlen(oscData.path)+13];
     std::strcpy(targetPath, oscData.path);
     std::strcat(targetPath, "/sample-rate");
-    try_lo_send(oscData.target, targetPath, "f", sampleRate);
+    try_lo_send(oscData.target, targetPath, "f", static_cast<double>(sampleRate));
 }
 
 static inline
